pyproject.toml
[tool.poetry]
name = "etf-trend-algo"
version = "0.1.0"
description = ""
authors = ["Moses McCall <mosesmc52@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
alpaca-trade-api = "^3.2.0"
boto3 = "^1.34.149"
sqlalchemy = "1.3.19"
pandas = "1.5.3"
xtermcolor = "^1.3"
fredapi = "^0.5.2"
sentry-sdk = "^2.11.0"
numpy = "1.26.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
